damn conversions

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -1,6 +1,7 @@
 #include "main.h"
 
-/* print_binary - Prints binary representation of a number
+/**
+ * print_binary - Prints binary representation of a number
  * @n: number to be printed
  *
  * Return: 0
@@ -9,10 +10,10 @@
 void print_binary(unsigned long int n)
 {
 	int c;
-	
+
 	tally = 0;
 
-	unsigned long int num;
+	unsigned int num;
 
 	for (c = 63; c >= 0; c--)
 	{
